Move crop file to server

+++ Source/Coco.Server/Coco.Entities/Model/General/UpdateAvatarModel.cs
@@ -3,10 +3,10 @@
     public class UpdateAvatarModel
     {
         public string PhotoUrl { get; set; }
-        public int XAxis { get; set; }
-        public int YAxis { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double XAxis { get; set; }
+        public double YAxis { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
         public string ContentType { get; set; }
         public bool CanEdit { get; set; }
     }
